+++ web/src/components/molecules/ProjectList/ProjectCard.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 
 import Card from "@reearth-cms/components/atoms/Card";
 import Icon from "@reearth-cms/components/atoms/Icon";
-import { Project } from "@reearth-cms/components/molecules/Dashboard/types";
+import { Project } from "@reearth-cms/components/molecules/Workspace/types";
 
 export interface Props {
   className?: string;
