+++ web/src/components/molecules/ProjectList/ProjectList.tsx
@@ -3,8 +3,8 @@ import React from "react";
 
 import Button from "@reearth-cms/components/atoms/Button";
 import Icon from "@reearth-cms/components/atoms/Icon";
-import ProjectCard from "@reearth-cms/components/molecules/Dashboard/ProjectCard";
-import { Project as ProjectType } from "@reearth-cms/components/molecules/Dashboard/types";
+import ProjectCard from "@reearth-cms/components/molecules/ProjectList/ProjectCard";
+import { Project as ProjectType } from "@reearth-cms/components/molecules/Workspace/types";
 import { useT } from "@reearth-cms/i18n";
 
 export interface Props {
@@ -14,7 +14,7 @@ export interface Props {
   handleProjectModalOpen: () => void;
 }
 
-const MoleculesProjectList: React.FC<Props> = ({
+const ProjectList: React.FC<Props> = ({
   className,
   projects,
   workspaceId,
@@ -92,4 +92,4 @@ const Content = styled.div`
   align-content: flex-start;
 `;
 
-export default MoleculesProjectList;
+export default ProjectList;
