+++ web/src/components/organisms/Workspace/hooks.ts
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from "react";
 
-import { Project } from "@reearth-cms/components/molecules/Dashboard/types";
+import { Project } from "@reearth-cms/components/molecules/Workspace/types";
 import { useGetProjectsQuery, useCreateProjectMutation } from "@reearth-cms/gql/graphql-client-api";
 import { useWorkspace } from "@reearth-cms/state";
